Lazy load App::CPANTS::Lint
Because all DZAC plugins are loaded for any command run (ex: kwalitee
will be loaded when we run just "dzil build"), each plugin must be light
to not impact the global performance of dzil.
The fix is to not load dependencies at compile time, but when needed, in
the 'execute' method.